last of clang

+++ src/video_utils.cpp
@@ -185,7 +185,7 @@ VideoFrame fromOwnedProto(const proto::OwnedVideoBuffer &owned) {
     throw std::runtime_error("fromOwnedProto: info.data_ptr is null");
   }
   const auto *src = reinterpret_cast<const std::uint8_t *>(
-      src_ptr); // NOLINT(performance-no-int-to-ptr)
+      static_cast<std::uintptr_t>(src_ptr));
 
   std::memcpy(dst, src, dst_size);
